+++ packages/server/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @simple-builder/server
 
+## 1.0.8
+
+### Patch Changes
+
+- fix bug with styles
+
 ## 1.0.7
 
 ### Patch Changes